Changed to socials and refactor
I changed the socials list to a flex box instead of handcrafted. Also, the socials section was weirdly position so it was resulting in bugs.
I refactored some of the codebase to get rid of a lot of repeated code. This was primarily in the socials section.
Another feature I added was making a max width and min width for the headers and changed the font size to .em instead of pixeled based. This fixed the error of text falling out of the box when you made the screen too small.